move createInitialState to testhelpers
